chore: update flag documentation
